+++ app/src/main/java/com/propertio/developer/unit/form/type/UnitDataApartemenFragment.kt
@@ -1,0 +1,35 @@
+package com.propertio.developer.unit.form.type
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.propertio.developer.R
+
+
+/**
+ * A simple [Fragment] subclass.
+ * Use the [UnitDataApartemenFragment.newInstance] factory method to
+ * create an instance of this fragment.
+ */
+class UnitDataApartemenFragment : Fragment() {
+    // TODO: Rename and change types of parameters
+    private var param1: String? = null
+    private var param2: String? = null
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+    }
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_unit_data_apartemen, container, false)
+    }
+
+
+}
