+++ app/src/main/java/com/propertio/developer/unit/form/type/UnitDataKondominiumFragment.kt
@@ -1,0 +1,28 @@
+package com.propertio.developer.unit.form.type
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.propertio.developer.R
+
+
+class UnitDataKondominiumFragment : Fragment() {
+
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+    }
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_unit_data_kondominium, container, false)
+    }
+
+
+}
